Fixes for GFortran 12
There seems to be a change in behavior between GFortran 11 and 12 regarding how allocatable character arrays are handled. GFortran < 12 seems to behave differently than newer versions. I see no way to make one Fortran code work with all, and the GFortran >= 12 seems to behave better and be compatible with Intel. Therefore choosing to use some preprocessor macros for this purpose. In the end we phase out GFortran 11 support.

Ref:
https://fortran-lang.discourse.group/t/allocatable-character-array-in-bind-c-dummy-argument-gfortran-11-bug/6546