Added integer field support
Support for integer fields of type intfield_t alongside field_t. Basic operations such as IO and connect are supported.

Integer fields are of basic kind ifk (Integer Field Kind). Default to 8 byte integers to trigger errors in building and testing if 'intk' and 'ifk' are mixed up. The rationale for creating a new kind ifk is that intk is used for nearly every integer in MGLET, kk, jj, ii etc. Setting 'intk' to any other value than int32 is failing and I doubt that it will ever work. Therefore it is easier to use a new kind and make this work properly.

In the end 'intk' could possibly be removed all together in favor of "default integers".