Probes 'maskbp' feature
Brobe arrays have a 'maskbp' flag, by default set. When this flag is set,
a check is performed on the BP field to see if it is possible to perform a
valid sample from the probe. When this is not possible, the grid number of
the probe is set to zero and the probe is disabled in the same
way as if it was outside the domain.