LLVM flang-new changes and additions
With these changes MGLET build and runs all basic testcases with LLVM flang-new, clang and clang++ compilers. No advanced testcases tried so far, though.

Be aware that flang is in rapid development, and this was tested with the latest main LLVM branch as of today. Older releases are likely to fail.