Tscastencil bugfix and cleanup
In the case that ngeomnbr == 0 - the variable body was never set to zero and would have the value from some random previous cell. This means that found2 was still set to 1 despite nothing was found.

In the case of ngeomnbr > 0 body was appropriately set to zero in line 211 - but that did not happen when ngeomnbr == 0.